Tables in cache post: align *and* center numbers
Right-align numbers in the tables of the post about CPU caches (my mom
said I should do that), but still center them based on the longest
number in a column.  This is accomplished by manually adding figure
spaces (spaces having the same width as digits [1]): a poor man's
version of the [siunitx] package's `S` column type (which I used for the
PDF).  This hack will break in case the font used has digits with
non-uniform widths.

See the accepted answer to <https://stackoverflow.com/q/13992026>.

[1]: https://en.wikipedia.org/wiki/Figure_space
[2]: https://ctan.org/pkg/siunitx