Use `px` instead of `mm`
This fixes issue #7 (images that are supposed to extend all the way to
the right don't on Safari).

It seems that Safari's `calc` rounds differently than other web browsers
and that we can work around this by using `px` instead of `mm` and by
using relatively even values.

Note that `px` is an absolute unit just like `mm` and that it has little
to do with display pixels.

https://github.com/meribold/meribold.org/issues/7
https://user-images.githubusercontent.com/3041190/125051438-8a6d3200-e0d5-11eb-938d-85eda4a6b1b4.png
https://user-images.githubusercontent.com/3041190/125051472-935e0380-e0d5-11eb-870b-69b349eb1e08.png
https://www.w3schools.com/cssref/css_units.asp
https://stackoverflow.com/q/16839847