Inline all the CSS and JavaScript
Inlining all the CSS and JS should speed up page loads for the typical
visitor.  The amount of data that needs to be loaded when someone visits
my site for the first time (or revisits after a while) remains
effectively the same, but everything is sent immediately without
additional round trips.  If a visitor clicks around on my website the
same CSS and JS has to be sent over and over again now since nothing can
be cached, but in practice a few extra kilobytes will hardly matter.  On
top of that, the typical visitor probably accesses one of my articles
and never navigates to any other page of my site.

https://developers.google.com/speed/pagespeed/insights/
https://web.dev/critical-request-chains/
https://web.dev/render-blocking-resources/
https://web.dev/optimize-lcp/#render-blocking-resources
https://web.dev/optimize-lcp/#inline-critical-css
https://blog.logrocket.com/improve-site-performance-inlining-css/
https://jekyllrb.com/docs/includes/
https://shopify.github.io/liquid/tags/template/