Virtual env. post: save hundreds of kilobytes
Serve the Russel photo in the WebP format [1] if the browser supports
it.  Serve the photo in a size that depends on the width and the height
in device pixels that the user will actually see.

The WebP images images were created with version 2.10.28 of the GIMP.
Specifically, version 2.10.28-1 of the Arch Linux package [2] was used.
Each time I scaled the PNG from [3] using cubic interpolation and
exported in the WebP format using an image quality setting of 90 out of
100 and otherwise GIMP's defaults.

Note that the AVIF format [4] may beat WebP but isn't widely supported
yet [5].

[1]: https://en.wikipedia.org/wiki/WebP
[2]: https://archlinux.org/packages/extra/x86_64/gimp/
[3]: https://upload.wikimedia.org/wikipedia/commons/archive/a/a6/20180321141731%21Bertrand_Russell_transparent_bg.png
[4]: https://en.wikipedia.org/wiki/AV1#AV1_Image_File_Format_(AVIF)
[5]: https://caniuse.com/avif

More links:

https://commons.wikimedia.org/wiki/File:Bertrand_Russell_transparent_bg.png
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/picture
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img#attr-sizes
https://developer.mozilla.org/en-US/docs/Web/Media/Formats/Image_types
https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Responsive_images
https://developers.google.com/speed/pagespeed/insights/
https://web.dev/uses-responsive-images/
https://web.dev/uses-webp-images/
https://web.dev/serve-images-with-correct-dimensions/
https://web.dev/optimize-cls/
https://web.dev/serve-responsive-images/
https://caniuse.com/webp
https://caniuse.com/srcset
https://caniuse.com/picture
https://css-tricks.com/responsive-images-youre-just-changing-resolutions-use-srcset/ (2015)