Virtual env. post: save dozens of kilobytes
Serve the Virtual Boy photo in the WebP format [1] if the browser
supports it.  Serve the photo in a size that depends on the width and
the height in device pixels that the user will actually see.

The WebP images images were created with version 2.10.28 of the GIMP.
Specifically, version 2.10.28-1 of the Arch Linux package [2] was used.
Each time I scaled the JPEG [3] using cubic interpolation and exported
in the WebP format using an image quality setting of 90 out of
100 and otherwise GIMP's defaults.

Also see commit eb4522c15eb4fb496b3368b749c2ac25a22abc2e.

[1]: https://en.wikipedia.org/wiki/WebP
[2]: https://archlinux.org/packages/extra/x86_64/gimp/
[3]: assets/virtual-boy-avgn.jpg

More links:

https://developers.google.com/speed/pagespeed/insights/
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/source
https://caniuse.com/webp