About page: embed photo of myself more skillfully
Serve the photo of myself in the WebP format if the browser supports it
and serve a size that depends on the width and the height in device
pixels that the user will actually see.

The WebP images images were created with version 2.10.28 of the GIMP.
Specifically, version 2.10.28-1 of the Arch Linux package was used.
Each time I scaled the JPEG using cubic interpolation and exported in
the WebP format using an image quality setting of 90 out of 100 and
otherwise GIMP's defaults.  This time I'm using git-annex for the new
images.

See commit 5a98bbf508ce0009611804cadc7 and eb4522c15eb4fb496b3368b749c.

Also prevent layout shift.  See commit e20f11cccae768d543fc333e74721fd9.