Add two WebP versions of the SnapScore screenshot
Serve the image in the WebP format if the browser supports it and serve
a size that depends on the width and the height in device pixels that
the user will actually see.

The WebP images images were created with version 2.10.28 of the GIMP.
Specifically, version 2.10.28-1 of the Arch Linux package was used.
Each time I scaled the PNG using cubic interpolation and exported in the
WebP format using an image quality setting of 90 out of 100 and
otherwise GIMP's defaults.  I'm using git-annex for the new images.

Also see these commits:

* eeb2c5b434f591af770a756111c18636c1095bd5
* 5a98bbf508ce0009611804cadc79c45d8ef18dd7
* eb4522c15eb4fb496b3368b749c2ac25a22abc2e