Ubuntu Mono is too special
Ubuntu Mono looks good when rendered at the same font size as
surrounding text that uses a proportional font.  But this is unusual:
most monospaced fonts need to be rendered at a smaller size.

We have to pick a size without knowing what font is going to be used, so
we have three options:

1. Pick a size that looks good with Ubuntu Mono and other monospaced
   fonts that are special in the same way (Inconsolata is another
   example).
2. Pick a size that looks good with most monospaced fonts other than
   Ubuntu Mono.
3. Pick a size that's somewhere in the middle.

This commit abandons option 3 in favor of option 2.